requirements/cdk.txt
aws-cdk.core==1.183.0
aws-cdk.aws-imagebuilder==1.183.0
aws-cdk.aws-iam==1.183.0
aws-cdk.aws-ssm==1.183.0
aws-cdk.aws-autoscaling==1.183.0
aws-cdk.aws-ec2==1.183.0
aws-cdk.aws-cloudfront==1.183.0
aws-cdk.aws-cloudfront-origins==1.183.0
aws-cdk.aws-codedeploy==1.183.0
aws-cdk.aws_certificatemanager==1.183.0
aws-cdk.aws-route53==1.183.0
aws-cdk.aws-route53-targets==1.183.0
aws-cdk.aws-lambda-python==1.183.0
aws-cdk.aws-events-targets==1.183.0
pyyaml==6.0
diagrams==0.23.1